requirements.txt
aiofiles==0.7.0
DateTime==4.3
discord==1.7.3
html2markdown==0.1.7
PyYAML==5.4.1
pyz3r==5.5.2
requests==2.26.0
tenacity==8.0.1
